+++ jobsdb/jobsdb.go
@@ -75,6 +75,14 @@ type JobsDB interface {
 }
 
 /*
+AssertInterface contains public assert methods
+*/
+type AssertInterface interface {
+	assert(cond bool, errorString string)
+	assertError(err error)
+}
+
+/*
 JobStatusT is used for storing status of the job. It is
 the responsibility of the user of this module to set appropriate
 job status. State can be one of
@@ -329,18 +337,6 @@ func getValidNonTerminalStates() (validNonTerminalStates []string) {
 	return
 }
 
-func (jd *HandleT) checkValidJobState(stateFilters []string) {
-	jobStateMap := make(map[string]jobStateT)
-	for _, js := range jobStates {
-		jobStateMap[js.State] = js
-	}
-	for _, st := range stateFilters {
-		js, ok := jobStateMap[st]
-		jd.assert(ok, fmt.Sprintf("state %s is not found in jobStates: %v", st, jobStates))
-		jd.assert(js.isValid, fmt.Sprintf("jobState : %v is not valid", js))
-	}
-}
-
 var (
 	host, user, password, dbname, sslmode string
 	port                                  int
@@ -549,91 +545,6 @@ func (jd *HandleT) TearDown() {
 	jd.dbHandle.Close()
 }
 
-/*
-Function to sort table suffixes. We should not have any use case
-for having > 2 len suffixes (e.g. 1_1_1 - see comment below)
-but this sort handles the general case
-*/
-func (jd *HandleT) sortDnumList(dnumList []string) {
-	sort.Slice(dnumList, func(i, j int) bool {
-		src := strings.Split(dnumList[i], "_")
-		dst := strings.Split(dnumList[j], "_")
-		comparison, err := dsComparitor(src, dst)
-		jd.assertError(err)
-		return comparison
-	})
-}
-
-// returns true, nil if src is less than dst
-// returns false, nil if src is greater than dst
-// returns false, someError if src is equal to dst
-var dsComparitor = func(src, dst []string) (bool, error) {
-	k := 0
-	for {
-		if k >= len(src) {
-			//src has same prefix but is shorter
-			//For example, src=1.1 while dest=1.1.1
-			if k >= len(dst) {
-				return false, fmt.Errorf("k:%d >= len(dst):%d", k, len(dst))
-			}
-			if k <= 0 {
-				return false, fmt.Errorf("k:%d <= 0", k)
-			}
-			return true, nil
-		}
-		if k >= len(dst) {
-			//Opposite of case above
-			if k <= 0 {
-				return false, fmt.Errorf("k:%d <= 0", k)
-			}
-			if k >= len(src) {
-				return false, fmt.Errorf("k:%d >= len(src):%d", k, len(src))
-			}
-			return false, nil
-		}
-		if src[k] == dst[k] {
-			//Loop
-			k++
-			continue
-		}
-		//Strictly ordered. Return
-		srcInt, err := strconv.Atoi(src[k])
-		if err != nil {
-			return false, fmt.Errorf("string to int conversion failed for source %v. with error %w", src, err)
-		}
-		dstInt, err := strconv.Atoi(dst[k])
-		if err != nil {
-			return false, fmt.Errorf("string to int conversion failed for destination %v. with error %w", dst, err)
-		}
-		return srcInt < dstInt, nil
-	}
-}
-
-//Function to get all table names form Postgres
-func (jd *HandleT) getAllTableNames() []string {
-	//Read the table names from PG
-	stmt, err := jd.dbHandle.Prepare(`SELECT tablename
-                                        FROM pg_catalog.pg_tables
-                                        WHERE schemaname != 'pg_catalog' AND
-                                        schemaname != 'information_schema'`)
-	jd.assertError(err)
-	defer stmt.Close()
-
-	rows, err := stmt.Query()
-	jd.assertError(err)
-	defer rows.Close()
-
-	tableNames := []string{}
-	for rows.Next() {
-		var tbName string
-		err = rows.Scan(&tbName)
-		jd.assertError(err)
-		tableNames = append(tableNames, tbName)
-	}
-
-	return tableNames
-}
-
 //removeExtraKey : removes extra key present in map1 and not in map2
 //Assumption is keys in map1 and map2 are same, except that map1 has one key more than map2
 func removeExtraKey(map1, map2 map[string]string) string {
@@ -673,62 +584,7 @@ func (jd *HandleT) getDSList(refreshFromDB bool) []dataSetT {
 	//Reset the global list
 	jd.datasetList = nil
 
-	//Read the table names from PG
-	tableNames := jd.getAllTableNames()
-
-	//Tables are of form jobs_ and job_status_. Iterate
-	//through them and sort them to produce and
-	//ordered list of datasets
-
-	jobNameMap := map[string]string{}
-	jobStatusNameMap := map[string]string{}
-	dnumList := []string{}
-
-	for _, t := range tableNames {
-		if strings.HasPrefix(t, jd.tablePrefix+"_jobs_") {
-			dnum := t[len(jd.tablePrefix+"_jobs_"):]
-			jobNameMap[dnum] = t
-			dnumList = append(dnumList, dnum)
-			continue
-		}
-		if strings.HasPrefix(t, jd.tablePrefix+"_job_status_") {
-			dnum := t[len(jd.tablePrefix+"_job_status_"):]
-			jobStatusNameMap[dnum] = t
-			continue
-		}
-	}
-
-	jd.sortDnumList(dnumList)
-
-	//If any service has crashed while creating DS, this may happen. Handling such case gracefully.
-	if len(jobNameMap) != len(jobStatusNameMap) {
-		jd.assert(len(jobNameMap) == len(jobStatusNameMap)+1, fmt.Sprintf("Length of jobNameMap(%d) - length of jobStatusNameMap(%d) is more than 1", len(jobNameMap), len(jobStatusNameMap)))
-		deletedDNum := removeExtraKey(jobNameMap, jobStatusNameMap)
-		//remove deletedDNum from dnumList
-		var idx int
-		var dnum string
-		var foundDeletedDNum bool
-		for idx, dnum = range dnumList {
-			if dnum == deletedDNum {
-				foundDeletedDNum = true
-				break
-			}
-		}
-		if foundDeletedDNum {
-			dnumList = remove(dnumList, idx)
-		}
-	}
-
-	//Create the structure
-	for _, dnum := range dnumList {
-		jobName, ok := jobNameMap[dnum]
-		jd.assert(ok, fmt.Sprintf("dnum %s is not found in jobNameMap", dnum))
-		jobStatusName, ok := jobStatusNameMap[dnum]
-		jd.assert(ok, fmt.Sprintf("dnum %s is not found in jobStatusNameMap", dnum))
-		jd.datasetList = append(jd.datasetList,
-			dataSetT{JobTable: jobName,
-				JobStatusTable: jobStatusName, Index: dnum})
-	}
+	jd.datasetList = getDSList(jd, jd.dbHandle, jd.tablePrefix)
 
 	//if the owner of this jobsdb is a writer, then shrinking datasetList to have only last two datasets
 	//this shrinked datasetList is used to compute DSRangeList
@@ -1324,7 +1180,7 @@ func (jd *HandleT) terminateQueries() {
 
 func (jd *HandleT) getBackupDSList() []dataSetT {
 	//Read the table names from PG
-	tableNames := jd.getAllTableNames()
+	tableNames := getAllTableNames(jd, jd.dbHandle)
 
 	jobNameMap := map[string]string{}
 	jobStatusNameMap := map[string]string{}
@@ -1366,7 +1222,7 @@ func (jd *HandleT) dropAllBackupDS() error {
 }
 
 func (jd *HandleT) dropMigrationCheckpointTables() {
-	tableNames := jd.getAllTableNames()
+	tableNames := getAllTableNames(jd, jd.dbHandle)
 
 	var migrationCheckPointTables []string
 	for _, t := range tableNames {
@@ -1568,33 +1424,6 @@ func (jd *HandleT) storeJobDS(ds dataSetT, job *JobT) (err error) {
 	return
 }
 
-func (jd *HandleT) constructQuery(paramKey string, paramList []string, queryType string) string {
-	jd.assert(queryType == "OR" || queryType == "AND", fmt.Sprintf("queryType:%s is neither OR nor AND", queryType))
-	var queryList []string
-	for _, p := range paramList {
-		queryList = append(queryList, "("+paramKey+"='"+p+"')")
-	}
-	return "(" + strings.Join(queryList, " "+queryType+" ") + ")"
-}
-
-func (jd *HandleT) constructParameterJSONQuery(table string, parameterFilters []ParameterFilterT) string {
-	// eg. query with optional destination_id (batch_rt_jobs_1.parameters @> '{"source_id":"<source_id>","destination_id":"<destination_id>"}'  OR (batch_rt_jobs_1.parameters @> '{"source_id":"<source_id>"}' AND batch_rt_jobs_1.parameters -> 'destination_id' IS NULL))
-	var allKeyValues, mandatoryKeyValues, opNullConditions []string
-	for _, parameter := range parameterFilters {
-		allKeyValues = append(allKeyValues, fmt.Sprintf(`"%s":"%s"`, parameter.Name, parameter.Value))
-		if parameter.Optional {
-			opNullConditions = append(opNullConditions, fmt.Sprintf(`%s.parameters -> '%s' IS NULL`, table, parameter.Name))
-		} else {
-			mandatoryKeyValues = append(mandatoryKeyValues, fmt.Sprintf(`"%s":"%s"`, parameter.Name, parameter.Value))
-		}
-	}
-	opQuery := ""
-	if len(opNullConditions) > 0 {
-		opQuery += fmt.Sprintf(` OR (%s.parameters @> '{%s}' AND %s)`, table, strings.Join(mandatoryKeyValues, ","), strings.Join(opNullConditions, " AND "))
-	}
-	return fmt.Sprintf(`(%s.parameters @> '{%s}' %s)`, table, strings.Join(allKeyValues, ","), opQuery)
-}
-
 type cacheValue string
 
 const (
@@ -1723,7 +1552,7 @@ parameterFilters do a AND query on values included in the map
 */
 func (jd *HandleT) getProcessedJobsDS(ds dataSetT, getAll bool, stateFilters []string,
 	customValFilters []string, limitCount int, parameterFilters []ParameterFilterT) []*JobT {
-	jd.checkValidJobState(stateFilters)
+	checkValidJobState(jd, stateFilters)
 
 	if jd.isEmptyResult(ds, stateFilters, customValFilters, parameterFilters) {
 		jd.logger.Debugf("[getProcessedJobsDS] Empty cache hit for ds: %v, stateFilters: %v, customValFilters: %v, parameterFilters: %v", ds, stateFilters, customValFilters, parameterFilters)
@@ -1748,14 +1577,14 @@ func (jd *HandleT) getProcessedJobsDS(ds dataSetT, getAll bool, stateFilters []s
 	var stateQuery, customValQuery, limitQuery, sourceQuery string
 
 	if len(stateFilters) > 0 {
-		stateQuery = " AND " + jd.constructQuery("job_state", stateFilters, "OR")
+		stateQuery = " AND " + constructQuery(jd, "job_state", stateFilters, "OR")
 	} else {
 		stateQuery = ""
 	}
 	if len(customValFilters) > 0 {
 		jd.assert(!getAll, "getAll is true")
 		customValQuery = " AND " +
-			jd.constructQuery(fmt.Sprintf("%s.custom_val", ds.JobTable),
+			constructQuery(jd, fmt.Sprintf("%s.custom_val", ds.JobTable),
 				customValFilters, "OR")
 	} else {
 		customValQuery = ""
@@ -1763,7 +1592,7 @@ func (jd *HandleT) getProcessedJobsDS(ds dataSetT, getAll bool, stateFilters []s
 
 	if len(parameterFilters) > 0 {
 		jd.assert(!getAll, "getAll is true")
-		sourceQuery += " AND " + jd.constructParameterJSONQuery(ds.JobTable, parameterFilters)
+		sourceQuery += " AND " + constructParameterJSONQuery(jd, ds.JobTable, parameterFilters)
 	} else {
 		sourceQuery = ""
 	}
@@ -1888,12 +1717,12 @@ func (jd *HandleT) getUnprocessedJobsDS(ds dataSetT, customValFilters []string,
 	}
 
 	if len(customValFilters) > 0 {
-		sqlStatement += " AND " + jd.constructQuery(fmt.Sprintf("%s.custom_val", ds.JobTable),
+		sqlStatement += " AND " + constructQuery(jd, fmt.Sprintf("%s.custom_val", ds.JobTable),
 			customValFilters, "OR")
 	}
 
 	if len(parameterFilters) > 0 {
-		sqlStatement += " AND " + jd.constructParameterJSONQuery(ds.JobTable, parameterFilters)
+		sqlStatement += " AND " + constructParameterJSONQuery(jd, ds.JobTable, parameterFilters)
 	}
 
 	if order {
@@ -2422,7 +2251,7 @@ func (jd *HandleT) getBackupDSRange() dataSetRangeT {
 	var backupDSRange dataSetRangeT
 
 	//Read the table names from PG
-	tableNames := jd.getAllTableNames()
+	tableNames := getAllTableNames(jd, jd.dbHandle)
 
 	//We check for job_status because that is renamed after job
 	dnumList := []string{}
@@ -2437,7 +2266,7 @@ func (jd *HandleT) getBackupDSRange() dataSetRangeT {
 		return backupDSRange
 	}
 
-	jd.sortDnumList(dnumList)
+	sortDnumList(jd, dnumList)
 
 	backupDS = dataSetT{
 		JobTable:       fmt.Sprintf("pre_drop_%s_jobs_%s", jd.tablePrefix, dnumList[0]),
@@ -2975,7 +2804,7 @@ parameterFilters do a AND query on values included in the map
 */
 func (jd *HandleT) deleteJobStatusDSInTxn(txHandler transactionHandler, ds dataSetT, stateFilters []string,
 	customValFilters []string, parameterFilters []ParameterFilterT) (int, error) {
-	jd.checkValidJobState(stateFilters)
+	checkValidJobState(jd, stateFilters)
 
 	var queryStat stats.RudderStats
 	statName := ""
@@ -2992,13 +2821,13 @@ func (jd *HandleT) deleteJobStatusDSInTxn(txHandler transactionHandler, ds dataS
 	var stateQuery, customValQuery, sourceQuery string
 
 	if len(stateFilters) > 0 {
-		stateQuery = " AND " + jd.constructQuery("job_state", stateFilters, "OR")
+		stateQuery = " AND " + constructQuery(jd, "job_state", stateFilters, "OR")
 	} else {
 		stateQuery = ""
 	}
 	if len(customValFilters) > 0 {
 		customValQuery = " WHERE " +
-			jd.constructQuery(fmt.Sprintf("%s.custom_val", ds.JobTable),
+			constructQuery(jd, fmt.Sprintf("%s.custom_val", ds.JobTable),
 				customValFilters, "OR")
 	} else {
 		customValQuery = ""
@@ -3011,7 +2840,7 @@ func (jd *HandleT) deleteJobStatusDSInTxn(txHandler transactionHandler, ds dataS
 	}
 
 	if len(parameterFilters) > 0 {
-		sourceQuery += jd.constructParameterJSONQuery(ds.JobTable, parameterFilters)
+		sourceQuery += constructParameterJSONQuery(jd, ds.JobTable, parameterFilters)
 	} else {
 		sourceQuery = ""
 	}
